FNDCongratsBackend/requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
boto3==1.24.26
botocore==1.27.26
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.0
Click==7.0
clipboard==0.0.4
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
discord==1.7.3
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
ec2-hibinit-agent==1.0.0
entrypoints==0.3
frozenlist==1.3.0
gql==3.3.0
graphql-core==3.2.1
h11==0.13.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==6.0.2
netifaces==0.10.4
numpy==1.23.1
oauthlib==3.1.0
outcome==1.2.0
pandas==1.4.3
pexpect==4.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyperclip==1.8.2
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.3.1
requests==2.28.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
s3transfer==0.6.0
SecretStorage==2.3.1
selenium==4.3.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
trio==0.21.0
trio-websocket==0.9.2
Twisted==18.9.0
typing-extensions==4.3.0
ubuntu-advantage-tools==27.8
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.10
wadllib==1.3.3
websockets==10.3
wsproto==1.1.0
yarl==1.7.2
zipp==1.0.0
zope.interface==4.7.1
